+++ public/index.html
@@ -9,6 +9,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
   <meta name="theme-color" content="#2ea3d2" />
   <meta name="description" content="Pictogram viewer and configurator" />
+  <style href="kits.css" />
 
   <link rel="apple-touch-icon" href="%PUBLIC_URL%/assets/pictosearch-icon-192.png" />
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
