+++ public/kits.css
@@ -1,0 +1,3 @@
+.test {
+  float: left;
+}
